+++ types/src/main/js/wallet/model/MoneroDestination.d.ts
@@ -7,9 +7,9 @@ declare class MoneroDestination {
      * Construct the model.
      *
      * @param {MoneroDestination|object|string} stateOrAddress is a MoneroDestination, JS object, or hex string to initialize from (optional)
-     * @param {BigInteger|string} amount - the destination amount
+     * @param {BigInt|string} amount - the destination amount
      */
-    constructor(stateOrAddress: MoneroDestination | object | string, amount: BigInteger | string);
+    constructor(stateOrAddress: MoneroDestination | object | string, amount: BigInt | string);
     state: any;
     getAddress(): any;
     setAddress(address: any): MoneroDestination;
@@ -19,5 +19,3 @@ declare class MoneroDestination {
     toJson(): any;
     toString(indent?: number): string;
 }
-import BigInteger_1 = require("../../common/biginteger");
-import BigInteger = BigInteger_1.BigInteger;
