+++ types/src/main/js/wallet/model/MoneroTxConfig.d.ts
@@ -12,7 +12,7 @@ declare class MoneroTxConfig {
      * let config1 = new MoneroTxConfig({<br>
      * &nbsp;&nbsp; accountIndex: 0,<br>
      * &nbsp;&nbsp; address: "59aZULsUF3YN...",<br>
-     * &nbsp;&nbsp; amount: new BigInteger("500000"),<br>
+     * &nbsp;&nbsp; amount: new BigInt("500000"),<br>
      * &nbsp;&nbsp; priority: MoneroTxPriority.NORMAL,<br>
      * &nbsp;&nbsp; relay: true<br>
      * });<br><br>
@@ -20,7 +20,7 @@ declare class MoneroTxConfig {
      *
      * @param {MoneroTxConfig|object} [config] - configures the transaction to create (optional)
      * @param {string} config.address - single destination address
-     * @param {BigInteger} config.amount - single destination amount
+     * @param {BigInt} config.amount - single destination amount
      * @@param {number} config.accountIndex - source account index to transfer funds from
      * @@param {number} config.subaddressIndex - source subaddress index to transfer funds from
      * @param {int[]} config.subaddressIndices - source subaddress indices to transfer funds from
@@ -32,7 +32,7 @@ declare class MoneroTxConfig {
      * @param {string} config.note - transaction note saved locally with the wallet
      * @param {string} config.recipientName - recipient name saved locally with the wallet
      * @param {boolean} config.canSplit - allow funds to be transferred using multiple transactions
-     * @param {BigInteger} config.belowAmount - for sweep requests, include outputs below this amount when sweeping wallet, account, subaddress, or all unlocked funds
+     * @param {BigInt} config.belowAmount - for sweep requests, include outputs below this amount when sweeping wallet, account, subaddress, or all unlocked funds
      * @param {boolean} config.sweepEachSubaddress - for sweep requests, sweep each subaddress individually instead of together if true
      * @param {string} config.keyImage - key image to sweep (ignored except in sweepOutput() requests)
      */
@@ -56,16 +56,16 @@ declare class MoneroTxConfig {
     /**
      * Set the amount of a single-destination configuration.
      *
-     * @param {BigInteger|string} amount - the amount to set for the single destination
+     * @param {BigInt|string} amount - the amount to set for the single destination
      * @return {MoneroTxConfig} this configuration for chaining
      */
-    setAmount(amount: BigInteger | string): MoneroTxConfig;
+    setAmount(amount: BigInt | string): MoneroTxConfig;
     /**
      * Get the amount of a single-destination configuration.
      *
-     * @return {BigInteger} the amount of the single destination
+     * @return {BigInt} the amount of the single destination
      */
-    getAmount(): BigInteger;
+    getAmount(): BigInt;
     addDestination(destinationOrAddress: any, amount: any): any;
     getDestinations(): any;
     setDestinations(destinations: any, ...args: any[]): MoneroTxConfig;
@@ -108,5 +108,3 @@ declare class MoneroTxConfig {
      */
     setKeyImage(keyImage: string): MoneroTxConfig;
 }
-import BigInteger_1 = require("../../common/biginteger");
-import BigInteger = BigInteger_1.BigInteger;
